libs/runtime/sdk/python/pyproject.toml
[build-system]
requires = ["maturin>=1.7,<2.0"]
build-backend = "maturin"

[project]
name = "celerity-runtime-sdk"
requires-python = ">=3.13"
classifiers = [
    "Programming Language :: Rust",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
dynamic = ["version"]

[dependency-groups]
dev = [
    "pytest",
    "maturin>=1.7,<2.0",
    "requests",
    "types-requests",
    "websocket-client",
    "python-dotenv",
    "redis",
    "mypy>=1.19.1",
]

[tool.mypy]
python_version = "3.13"
warn_unused_configs = true
check_untyped_defs = true

[tool.maturin]
python-source = "python"
module-name = "celerity_runtime_sdk._celerity_runtime_sdk"
bindings = "pyo3"
features = ["pyo3/extension-module"]
